Adds a yelp client i.e. a retrofit client class - instance which uses the interface and the Builder API which allows defining the URL endpoint for the HTTP operation.

+++ app/src/main/java/com/example/myrestaurants/Business.java
@@ -97,11 +97,11 @@ public class Business {
         this.transactions = transactions;
     }
 
-    public Integer getRating() {
+    public Double getRating() {
         return rating;
     }
 
-    public void setRating(Integer rating) {
+    public void setRating(Double rating) {
         this.rating = rating;
     }
